+++ src/parsing/parse_utils.c
@@ -6,11 +6,14 @@
 /*   By: mattcarniel <mattcarniel@student.42.fr>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/02/12 12:25:46 by fadzejli          #+#    #+#             */
-/*   Updated: 2026/02/17 15:38:45 by mattcarniel      ###   ########.fr       */
+/*   Updated: 2026/02/18 00:10:54 by smamalig         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../../include/cub3d.h"
+#include "cub3d.h"
+#include "libft.h"
+#include <stddef.h>
+#include <stdint.h>
 
 int	find_type(char *line)
 {
@@ -34,17 +37,17 @@ int	find_type(char *line)
 	return (0);
 }
 
-int	get_width(char **map)
+uint32_t	get_width(char **map)
 {
-	int	max_width;
-	int	width;
-	int	i;
+	uint32_t	max_width;
+	uint32_t	width;
+	uint32_t	i;
 
 	max_width = 0;
 	i = 0;
 	while (map[i])
 	{
-		width = ft_strlen(map[i]);
+		width = (uint32_t)ft_strlen(map[i]);
 		if (width > max_width)
 			max_width = width;
 		i++;
@@ -52,9 +55,9 @@ int	get_width(char **map)
 	return (max_width);
 }
 
-int	get_height(char **map)
+uint32_t	get_height(char **map)
 {
-	int	i;
+	uint32_t	i;
 
 	i = 0;
 	while (map[i])
@@ -83,9 +86,9 @@ int	check_empty_lines_after(int fd)
 
 char	**copy_map(char **map)
 {
-	char	**copy;
-	int		i;
-	int		height;
+	char		**copy;
+	uint32_t	i;
+	uint32_t	height;
 
 	height = get_height(map);
 	copy = malloc(sizeof(char *) * (height + 1));
