+++ src/threads/init.c
@@ -6,7 +6,7 @@
 /*   By: smamalig <smamalig@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/02/17 19:44:44 by smamalig          #+#    #+#             */
-/*   Updated: 2026/02/17 19:58:01 by smamalig         ###   ########.fr       */
+/*   Updated: 2026/02/18 00:25:45 by smamalig         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -47,7 +47,7 @@ int	threadpool_init(t_threadpool *pool, int worker_count)
 	int32_t	i;
 
 	ft_memset(pool, 0, sizeof(t_threadpool));
-	pool->threads = ft_calloc(worker_count, sizeof(pthread_t));
+	pool->threads = ft_calloc((size_t)worker_count, sizeof(pthread_t));
 	if (!pool->threads)
 		return (1);
 	pthread_mutex_init(&pool->lock, nullptr);
